add fixes to home.jsp

+++ src/main/webapp/js/script.js
@@ -6,6 +6,14 @@ $(function () {
         fullName: $("#requestUserFullName").text()
     };
 
+    var sessionUser = {
+        id: $("#sessionUserFirstName").data("user-id")
+    };
+    
+    function requestUserIsSessionUser(reqUser, sesUser) {
+        return sessionUser.id === reqUser.id;
+    }
+
     var postsContainer = $("#posts");
 
     var offsetPostId = 100000000;
@@ -42,11 +50,13 @@ $(function () {
                     prepareHtml += '<p>' + post.text + '</p>';
                     prepareHtml += '</div>';
 
-                    prepareHtml += '<div class="row">';
-                    prepareHtml += '<span class="post-remove glyphicon glyphicon-remove-circle"></span>';
-                    prepareHtml += '<span class="post-edit glyphicon glyphicon-edit"></span>';
-                    prepareHtml += '<span class="hidden post-edit-ok glyphicon glyphicon-ok"></span>';
-                    prepareHtml += '</div>';
+                    if (requestUserIsSessionUser(requestUser, sessionUser)) {
+                        prepareHtml += '<div class="row">';
+                        prepareHtml += '<span class="post-remove glyphicon glyphicon-remove-circle"></span>';
+                        prepareHtml += '<span class="post-edit glyphicon glyphicon-edit"></span>';
+                        prepareHtml += '<span class="hidden post-edit-ok glyphicon glyphicon-ok"></span>';
+                        prepareHtml += '</div>';
+                    }
 
                     prepareHtml += '</div>';
 
@@ -72,54 +82,56 @@ $(function () {
 
     loadPosts(requestUser, offsetPostId);
 
-    postsContainer.on('click', '.post-remove', function () {
-        var postId = $(this).closest('.post').data("post-id");
-
-        $.ajax({
-            url: "/restapi/posts/" + postId,
-            type: "DELETE",
-            dataType: "json",
-            success: $(this).closest('.post').remove()
-        });
-    });
-
-    postsContainer.on('click', '.post-edit', function () {
-        var p = $(this).parent('div').siblings('div').children('p');
-        var save = $(this).siblings('.post-edit-ok');
-        var edit = $(this);
-        var text = p.text().replace("\n", "").trim();
-
-        p.replaceWith("<textarea class='post-edit-text'>" + text + "</textarea>");
-
-        edit.addClass("hidden");
-        save.removeClass("hidden");
-
-        save.off('click').click(function () {
+    if (requestUserIsSessionUser(requestUser, sessionUser)) {
+        postsContainer.on('click', '.post-remove', function () {
             var postId = $(this).closest('.post').data("post-id");
-            var textArea = $(this).parent('div').siblings('div').children("textarea");
-            var updatedText = textArea.val();
-            var currentTime = new Date().toISOString();
 
             $.ajax({
-                url: "/restapi/posts",
-                type: "PUT",
-                method: "PUT",
-                headers: {'Content-Type' : 'application/json'},
+                url: "/restapi/posts/" + postId,
+                type: "DELETE",
                 dataType: "json",
-                data: JSON.stringify({
-                    id: postId,
-                    fromId: requestUser.id,
-                    postType: 0,
-                    text: updatedText,
-                    publishTime: currentTime
-                }),
-                success: function() {
-                    textArea.replaceWith('<p>' + updatedText + '</p>');
-
-                    save.addClass("hidden");
-                    edit.removeClass("hidden");
-                }
+                success: $(this).closest('.post').remove()
             });
         });
-    });
+
+        postsContainer.on('click', '.post-edit', function () {
+            var p = $(this).parent('div').siblings('div').children('p');
+            var save = $(this).siblings('.post-edit-ok');
+            var edit = $(this);
+            var text = p.text().replace("\n", "").trim();
+
+            p.replaceWith("<textarea class='post-edit-text'>" + text + "</textarea>");
+
+            edit.addClass("hidden");
+            save.removeClass("hidden");
+
+            save.off('click').click(function () {
+                var postId = $(this).closest('.post').data("post-id");
+                var textArea = $(this).parent('div').siblings('div').children("textarea");
+                var updatedText = textArea.val();
+                var currentTime = new Date().toISOString();
+
+                $.ajax({
+                    url: "/restapi/posts",
+                    type: "PUT",
+                    method: "PUT",
+                    headers: {'Content-Type': 'application/json'},
+                    dataType: "json",
+                    data: JSON.stringify({
+                        id: postId,
+                        fromId: requestUser.id,
+                        postType: 0,
+                        text: updatedText,
+                        publishTime: currentTime
+                    }),
+                    success: function () {
+                        textArea.replaceWith('<p>' + updatedText + '</p>');
+
+                        save.addClass("hidden");
+                        edit.removeClass("hidden");
+                    }
+                });
+            });
+        });
+    }
 });
